refactor: 워크스페이스 페이지에 있던 상태 Workspace 컴포넌트로 이동
- 해당 상태를 ConfMediaBar에서 사용하지 않음
- Workspace 컴포넌트는 회의록 서버, ConfMediaBar 컴포넌트는 시그널링 서버에 소켓 연결
